+++ example/lib/main.dart
@@ -28,24 +28,50 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
+  Future<List<CastDevice>>? _future;
+
   @override
   void initState() {
     super.initState();
-    CastDiscoveryService().start();
+    _startSearch();
   }
 
   @override
   Widget build(BuildContext context) {
-    return StreamBuilder<List<CastDevice>>(
-      stream: CastDiscoveryService().stream,
-      initialData: CastDiscoveryService().devices,
+    return FutureBuilder<List<CastDevice>>(
+      future: _future,
       builder: (context, snapshot) {
+        if (snapshot.hasError) {
+          return Center(
+            child: Text(
+              'Error: ${snapshot.error.toString()}',
+            ),
+          );
+        } else if (!snapshot.hasData) {
+          return Center(
+            child: CircularProgressIndicator(),
+          );
+        }
+
+        if (snapshot.data!.isEmpty) {
+          return Column(
+            children: [
+              Center(
+                child: Text(
+                  'No Chromecast founded',
+                ),
+              ),
+            ],
+          );
+        }
+
         return Column(
-          children: snapshot.data.map((device) {
+          children: snapshot.data!.map((device) {
             return ListTile(
               title: Text(device.name),
               onTap: () {
-                _connect(context, device);
+                // _connectToYourApp(context, device);
+                _connectAndPlayMedia(context, device);
               },
             );
           }).toList(),
@@ -54,7 +80,11 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 
-  Future<void> _connect(BuildContext context, CastDevice object) async {
+  void _startSearch() {
+    _future = CastDiscoveryService().search();
+  }
+
+  Future<void> _connectToYourApp(BuildContext context, CastDevice object) async {
     final session = await CastSessionManager().startSession(object);
 
     session.stateStream.listen((state) {
@@ -62,7 +92,7 @@ class _MyHomePageState extends State<MyHomePage> {
         final snackBar = SnackBar(content: Text('Connected'));
         Scaffold.of(context).showSnackBar(snackBar);
 
-        _sendMessage(session);
+        _sendMessageToYourApp(session);
       }
     });
 
@@ -72,13 +102,73 @@ class _MyHomePageState extends State<MyHomePage> {
 
     session.sendMessage(CastSession.kNamespaceReceiver, {
       'type': 'LAUNCH',
-      'appId': 'YouTube', // set the appId of your app here
+      'appId': 'Youtube', // set the appId of your app here
     });
   }
 
-  void _sendMessage(CastSession session) {
+  void _sendMessageToYourApp(CastSession session) {
+    print('_sendMessageToYourApp');
+
     session.sendMessage('urn:x-cast:namespace-of-the-app', {
       'type': 'sample',
+    });
+  }
+
+  Future<void> _connectAndPlayMedia(BuildContext context, CastDevice object) async {
+    final session = await CastSessionManager().startSession(object);
+
+    session.stateStream.listen((state) {
+      if (state == CastSessionState.connected) {
+        final snackBar = SnackBar(content: Text('Connected'));
+        Scaffold.of(context).showSnackBar(snackBar);
+      }
+    });
+
+    var index = 0;
+
+    session.messageStream.listen((message) {
+      index += 1;
+
+      print('receive message: $message');
+
+      if (index == 2) {
+        Future.delayed(Duration(seconds: 5)).then((x) {
+          _sendMessagePlayVideo(session);
+        });
+      }
+    });
+
+    session.sendMessage(CastSession.kNamespaceReceiver, {
+      'type': 'LAUNCH',
+      'appId': 'CC1AD845', // set the appId of your app here
+    });
+  }
+
+  void _sendMessagePlayVideo(CastSession session) {
+    print('_sendMessagePlayVideo');
+
+    var message = {
+      // Here you can plug an URL to any mp4, webm, mp3 or jpg file with the proper contentType.
+      'contentId': 'http://commondatastorage.googleapis.com/gtv-videos-bucket/big_buck_bunny_1080p.mp4',
+      'contentType': 'video/mp4',
+      'streamType': 'BUFFERED', // or LIVE
+
+      // Title and cover displayed while buffering
+      'metadata': {
+        'type': 0,
+        'metadataType': 0,
+        'title': "Big Buck Bunny",
+        'images': [
+          {'url': 'http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/images/BigBuckBunny.jpg'}
+        ]
+      }
+    };
+
+    session.sendMessage(CastSession.kNamespaceMedia, {
+      'type': 'LOAD',
+      'autoPlay': true,
+      'currentTime': 0,
+      'media': message,
     });
   }
 }