+++ lib/device.dart
@@ -7,10 +7,19 @@ class CastDevice {
   final String host;
   final int port;
 
+  final Map<String, String> extras;
+
   const CastDevice({
-    this.serviceName,
-    this.name,
-    this.host,
-    this.port,
+    required this.serviceName,
+    required this.name,
+    required this.host,
+    required this.port,
+    this.extras = const <String, String>{},
   });
+
+  @override
+  bool operator ==(Object other) => identical(this, other) || (other is CastDevice && runtimeType == other.runtimeType && other.serviceName == serviceName);
+
+  @override
+  int get hashCode => serviceName.hashCode;
 }