+++ lib/discovery_service.dart
@@ -1,8 +1,10 @@
 import 'dart:async';
 
-import 'package:mdns_plugin/mdns_plugin.dart';
+import 'package:multicast_dns/multicast_dns.dart';
 
 import 'device.dart';
+
+const _domain = '_googlecast._tcp.local';
 
 class CastDiscoveryService {
   static final CastDiscoveryService _instance = CastDiscoveryService._();
@@ -12,86 +14,72 @@ class CastDiscoveryService {
     return _instance;
   }
 
-  List<CastDevice> get devices => _delegate.devices ?? <CastDevice>[];
+  Future<List<CastDevice>> search({Duration timeout = const Duration(seconds: 5)}) async {
+    final results = <CastDevice>[];
+    final client = MDnsClient();
 
-  Stream<List<CastDevice>> get stream => _controller?.stream;
-  StreamController<List<CastDevice>> _controller;
+    await client.start();
 
-  MDNSPlugin _mdns;
-  _Delegate _delegate;
+    await for (PtrResourceRecord ptr in client.lookup<PtrResourceRecord>(ResourceRecordQuery.serverPointer(_domain), timeout: timeout)) {
+      if (results.any((x) => x.serviceName == ptr.domainName)) {
+        continue;
+      }
 
-  Future<void> start() async {
-    if (_mdns == null) {
-      _controller = StreamController<List<CastDevice>>.broadcast();
-      _delegate = _Delegate(_controller);
-      _mdns = MDNSPlugin(_delegate);
-      await _mdns.startDiscovery('_googlecast._tcp', enableUpdating: true);
+      await for (final srv in client.lookup<SrvResourceRecord>(ResourceRecordQuery.service(ptr.domainName))) {
+        final serviceName = ptr.domainName;
+        final port = srv.port;
+        String host = srv.target; // doesn't seem to work, => resolve IPv4
+        Map<String, String>? extras;
+
+        String? nameFromExtras;
+        final nameFromDomain = ptr.domainName.replaceAll('.$_domain', '').replaceAll(srv.target.replaceAll('.local', '').replaceAll('-', ''), '').replaceAll('-', ' ').trim();
+
+        await for (final ipAddress in client.lookup<IPAddressResourceRecord>(ResourceRecordQuery.addressIPv4(srv.target))) {
+          host = ipAddress.address.address;
+        }
+
+        if (host == srv.target) {
+          await for (final ipAddress in client.lookup<IPAddressResourceRecord>(ResourceRecordQuery.addressIPv6(srv.target))) {
+            host = ipAddress.address.address;
+          }
+        }
+
+        await for (final text in client.lookup<TxtResourceRecord>(ResourceRecordQuery.text(ptr.domainName))) {
+          extras = text.text.split('\n').fold<Map<String, String>>(<String, String>{}, (Map<String, String> acc, line) {
+            final values = line.split('=');
+
+            if (values.length < 2) {
+              return acc;
+            }
+
+            final key = values.first;
+            final value = values.skip(1).join('');
+
+            acc[key] = value;
+            return acc;
+          });
+
+          if (extras['fn']?.isNotEmpty ?? false) {
+            nameFromExtras = extras['fn']; // Chromecast function (Office, Living room...)
+          } else if (extras['md']?.isNotEmpty ?? false) {
+            nameFromExtras = extras['md']; // Chromecast model (Chromecast Ultra...)
+          }
+        }
+
+        final device = CastDevice(
+          serviceName: serviceName,
+          name: nameFromExtras ?? nameFromDomain,
+          host: host,
+          port: port,
+          extras: extras ?? {},
+        );
+
+        results.add(device);
+      }
     }
-  }
 
-  Future<void> stop() async {
-    await _mdns?.stopDiscovery();
-    await _controller?.close();
-    _controller = null;
-    _delegate = null;
-    _mdns = null;
-  }
-}
+    client.stop();
 
-class _Delegate implements MDNSPluginDelegate {
-  final devices = <CastDevice>[];
-  final StreamController<List<CastDevice>> controller;
-
-  _Delegate(this.controller);
-
-  @override
-  bool onServiceFound(MDNSService service) {
-    return true;
-  }
-
-  @override
-  void onServiceResolved(MDNSService service) {
-    devices.removeWhere((x) => x.serviceName == service.name);
-
-    final device = _transformServiceInDevice(service);
-    if (device != null) {
-      devices.add(device);
-    } else {
-      print('[CastDiscoveryService] onServiceResolved unable to read service: $service');
-    }
-
-    controller.sink.add(devices);
-  }
-
-  @override
-  void onServiceUpdated(MDNSService service) {}
-
-  @override
-  void onServiceRemoved(MDNSService service) {
-    devices.removeWhere((x) => x.serviceName == service.name);
-    controller.sink.add(devices);
-  }
-
-  @override
-  void onDiscoveryStarted() {}
-
-  @override
-  void onDiscoveryStopped() {}
-
-  static CastDevice _transformServiceInDevice(MDNSService service) {
-    String host = service.map['hostName'] ?? service.addresses.first;
-
-    String name;
-    if (service.txt != null && service.txt['fn'] != null) {
-      name = MDNSService.toUTF8String(service.txt['fn']);
-    }
-    name ??= service.name;
-
-    return CastDevice(
-      serviceName: service.name,
-      name: name,
-      host: host,
-      port: service.port,
-    );
+    return results.toSet().toList();
   }
 }